+++ src/server_frame/logic/player_manager.cpp
@@ -223,7 +223,8 @@ rpc::result_code_type player_manager::create(rpc::context &ctx, uint64_t user_id
     if (output->get_login_info().account().version_type() >= PROJECT_NAMESPACE_ID::EN_VERSION_INNER) {
       res = RPC_AWAIT_CODE_RESULT(output->create_init(ctx, PROJECT_NAMESPACE_ID::EN_VERSION_DEFAULT));
     } else {
-      res = RPC_AWAIT_CODE_RESULT(output->create_init(ctx, output->get_login_info().account().version_type()));
+      res = RPC_AWAIT_CODE_RESULT(
+          output->create_init(ctx, static_cast<uint32_t>(output->get_login_info().account().version_type())));
     }
     if (res < 0) {
       FWLOGERROR("save create {}:{} object failed, res: {}({})", zone_id, user_id, res,
